CS144 Project
A Web service and a Web site that allow users to search and navigate
eBay bidding data. The system manages all of its data at the back-end in
the MySQL database and supports browsing and searching interfaces over
the Internet. These functionalities are provided by integrating MySQL,
Apache Tomcat, Lucene and Axis2 packages.